pyproject.toml
[tool.poetry]
name = "lolli"
version = "0.1.0"
description = ""
authors = ["maccam912 <maccam912@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
typer = "^0.3.2"
Backtesting = "^0.3.1"
tda-api = "^1.3.2"
pandas = "^1.2.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.6b0"

[tool.poetry.scripts]
lolli = "lolli.run:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
